Create new tag for latest; push updated latest pointer to Quay
After a new Docker image is pushed to Quay (tagged with either a SHA or
a Git tag), create a new tag of that same reference, but name it
`latest`. Then, push that tag to Quay (shouldn't require much transfer
because all of the image layers should already be present at the
destination).